requirements.txt
#pip install -r .\requirements.txt 
gunicorn == 20.1.0
keras == 2.9.0
flask == 2.1.2
tensorflow-cpu == 2.9.1
Pillow == 9.1.1